+++ lib/shopify_api/rest/carrier_service.ex
@@ -12,9 +12,11 @@ defmodule ShopifyAPI.REST.CarrierService do
   ## Example
 
       iex> ShopifyAPI.REST.CarrierService.all(auth)
-      {:ok, { "carrier_services" => [] }}
+      {:ok, [%{}, ...] = carrier_services}
   """
-  def all(%AuthToken{} = auth, params \\ []), do: REST.get(auth, "carrier_services.json", params)
+  def all(%AuthToken{} = auth, params \\ [], options \\ []),
+    do:
+      REST.get(auth, "carrier_services.json", params, Keyword.merge([pagination: :none], options))
 
   @doc """
   Get a single carrier service.
@@ -22,10 +24,16 @@ defmodule ShopifyAPI.REST.CarrierService do
   ## Example
 
       iex> ShopifyAPI.REST.CarrierService.get(auth, string)
-      {:ok, { "carrier_service" => %{} }}
+      {:ok, %{} = carrier_service}
   """
-  def get(%AuthToken{} = auth, carrier_service_id, params \\ []),
-    do: REST.get(auth, "carrier_services/#{carrier_service_id}.json", params)
+  def get(%AuthToken{} = auth, carrier_service_id, params \\ [], options \\ []),
+    do:
+      REST.get(
+        auth,
+        "carrier_services/#{carrier_service_id}.json",
+        params,
+        Keyword.merge([pagination: :none], options)
+      )
 
   @doc """
   Create a carrier service.
@@ -33,7 +41,7 @@ defmodule ShopifyAPI.REST.CarrierService do
   ## Example
 
       iex> ShopifyAPI.REST.CarrierService.create(auth, map)
-      {:ok, { "carrier_service" => %{} }}
+      {:ok, %{} = carrier_service}
   """
   def create(%AuthToken{} = auth, %{carrier_service: %{}} = carrier_service),
     do: REST.post(auth, "carrier_services.json", carrier_service)
@@ -44,7 +52,7 @@ defmodule ShopifyAPI.REST.CarrierService do
   ## Example
 
       iex> ShopifyAPI.REST.CarrierService.update(auth)
-      {:ok, { "carrier_service" => %{} }}
+      {:ok, %{} = carrier_service}
   """
   def update(
         %AuthToken{} = auth,