+++ lib/shopify_api/rest/request.ex
@@ -12,29 +12,44 @@ defmodule ShopifyAPI.REST.Request do
   alias HTTPoison.Error
   alias ShopifyAPI.{AuthToken, CallLimit, JSONSerializer, Throttled}
 
-  @default_api_version "2019-04"
+  @default_api_version "2020-01"
 
   # Use HTTP in test for Bypass, HTTPS in all other environments
   @transport if Mix.env() == :test, do: "http://", else: "https://"
 
   @http_receive_timeout Application.get_env(:shopify_api, :http_timeout)
 
+  @type http_method :: :get | :post | :put | :delete
+
   ## Public Interface
 
+  @spec perform(
+          AuthToken.t(),
+          http_method(),
+          path :: String.t(),
+          body :: String.t(),
+          params :: keyword()
+        ) :: {:ok, HTTPoison.Response.t()} | {:error, HTTPoison.Response.t() | any()}
   def perform(%AuthToken{} = token, method, path, body \\ "", params \\ []) do
     url = token |> url(path) |> add_params_to_url(params)
     headers = headers(token)
 
-    response =
+    transform_response(
       Throttled.request(
         fn -> logged_request(method, url, body, headers, token: token) end,
         token
       )
+    )
+  end
 
+  defguard is_success_status(status) when status >= 200 and status < 300
+
+  @spec transform_response(HTTPoison.Response.t()) :: {:error, any()} | {:ok, any()}
+  def transform_response(response) do
     case response do
-      {:ok, %{status_code: status} = response} when status >= 200 and status < 300 ->
-        # TODO probably have to return the response here if we want to use the headers
-        {:ok, fetch_body(response)}
+      {:ok, %{status_code: status} = response}
+      when is_success_status(status) ->
+        {:ok, response}
 
       {:ok, response} ->
         {:error, response}
@@ -45,6 +60,36 @@ defmodule ShopifyAPI.REST.Request do
       response ->
         {:error, response}
     end
+  end
+
+  @spec stream(AuthToken.t(), String.t(), keyword()) :: Enumerable.t()
+  def stream(auth, path, params) do
+    headers = headers(auth)
+
+    start_fun = fn -> auth |> url(path) |> add_params_to_url(params) end
+
+    next_fun = fn
+      url when is_binary(url) ->
+        shopify_response =
+          Throttled.request(fn -> logged_request(:get, url, "", headers, token: auth) end, auth)
+
+        case transform_response(shopify_response) do
+          {:ok, resp} ->
+            extract_results_and_next_link(resp)
+
+          value ->
+            {[value], nil}
+        end
+
+      _ ->
+        {:halt, nil}
+    end
+
+    Stream.resource(
+      start_fun,
+      next_fun,
+      fn _ -> :ok end
+    )
   end
 
   def version do
@@ -74,7 +119,10 @@ defmodule ShopifyAPI.REST.Request do
 
   @impl true
   def process_response_body(body) do
-    JSONSerializer.decode(body)
+    with {:ok, results_map} when is_map(results_map) <- JSONSerializer.decode(body),
+         [{_key, results}] <- Map.to_list(results_map) do
+      results
+    end
   end
 
   ## Private Helpers
@@ -157,12 +205,6 @@ defmodule ShopifyAPI.REST.Request do
     ]
   end
 
-  defp fetch_body(http_response) do
-    with {:ok, map_fetched} <- Map.fetch(http_response, :body),
-         {:ok, body} <- map_fetched,
-         do: body
-  end
-
   @spec add_params_to_url(binary, list | map) :: binary
   defp add_params_to_url(url, params) when is_map(params) do
     list_params = Enum.map(params, fn {key, value} -> {String.to_existing_atom(key), value} end)
@@ -197,5 +239,26 @@ defmodule ShopifyAPI.REST.Request do
     for {key, value} <- list, into: Map.new() do
       {"#{key}", value}
     end
+  end
+
+  @link_regex ~r/<(?<link>.*)>;\s*rel=\"(?<rel>.*)\"/
+
+  @spec extract_next_link(list) :: binary | nil
+  defp extract_next_link(headers) do
+    for {"Link", link_header} <- headers,
+        links <- String.split(link_header, ",") do
+      case Regex.named_captures(@link_regex, links) do
+        %{"link" => link, "rel" => "next"} -> link
+        _ -> nil
+      end
+    end
+    |> Enum.filter(&(not is_nil(&1)))
+    |> List.first()
+  end
+
+  @spec extract_results_and_next_link(HTTPoison.Response.t()) :: {any(), nil | binary}
+  def extract_results_and_next_link(%HTTPoison.Response{body: body, headers: headers}) do
+    next_link = extract_next_link(headers)
+    {body, next_link}
   end
 end