+++ lib/shopify_api/rest/transaction.ex
@@ -7,8 +7,8 @@ defmodule ShopifyAPI.REST.Transaction do
   @doc """
     Return all the Transactions for an Order.
   """
-  def all(%AuthToken{} = auth, order_id, params \\ []),
-    do: REST.get(auth, "orders/#{order_id}/transactions.json", params)
+  def all(%AuthToken{} = auth, order_id, params \\ [], options \\ []),
+    do: REST.get(auth, "orders/#{order_id}/transactions.json", params, options)
 
   def create(%AuthToken{} = auth, %{transaction: %{order_id: order_id}} = transaction),
     do: REST.post(auth, "orders/#{order_id}/transactions.json", transaction)