+++ test/shopify_api/rest/product_test.exs
@@ -19,17 +19,14 @@ defmodule ShopifyAPI.REST.ProductTest do
   end
 
   test "", %{bypass: bypass, shop: _shop, auth_token: token} do
-    products = %{
-      "products" => [
-        %{"product_id" => "_", "title" => "Testing Create Product"}
-      ]
-    }
+    product = %{"product_id" => "_", "title" => "Testing Create Product"}
 
     Bypass.expect_once(bypass, "GET", "/admin/api/#{Request.version()}/products.json", fn conn ->
-      {:ok, body} = JSONSerializer.encode(products)
+      {:ok, body} = JSONSerializer.encode(%{products: [product]})
+
       Conn.resp(conn, 200, body)
     end)
 
-    assert {:ok, ^products} = Product.all(token)
+    assert {:ok, [^product]} = Product.all(token)
   end
 end